pyproject.toml
[project]
name = "ddd-building-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-storage-blob>=12.25.1",
    "fiona>=1.10.1",
    "gdal==3.7.3",
    "geopandas>=1.0.1",
    "hera[yaml]>=5.20.1",
    "laspy[laszip,lazrs]>=2.5.4",
    "mypy>=1.15.0",
    "pandas-stubs>=2.2.3.250308",
    "requests>=2.32.3",
    "shapely>=2.1.0",
    "tenacity>=9.1.2",
    "toml>=0.10.2",
    "tqdm>=4.67.1",
    "types-requests>=2.32.0.20250515",
    "types-shapely>=2.0.0.20250404",
    "types-toml>=0.10.8.20240310",
]
